// src/types/api_res/ProductDetail/ProductOption.ts
export interface ProductOption {
	eos_yn: string;
	stock_seq: number;
	stock: number;
	limit_cnt: number;
	soldout_yn: string;
	restock_date: null;
	stock_opt_price: number;
	opt_cd2: string;
	opt_nm2: string;
	opt_cd1: string;
	product_cd: string;
	opt_price: number;
	opt_price_bak: null;
	opt_color: string;
	use_yn: string;
	del_yn: string;
	od: number;
	opt_nm1: string;
	opt_gb: string;
	tot_combine_cnt: number;
	combine_idx: number;
}